SLING-5866 DefaultGetServlet obtains input stream for binary even if request is a HEAD
