+++ bundles/servlets/get/src/main/java/org/apache/sling/servlets/get/impl/helpers/StreamRendererServlet.java
@@ -50,6 +50,7 @@ import org.apache.sling.api.resource.ResourceMetadata;
 import org.apache.sling.api.resource.ResourceNotFoundException;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ResourceUtil;
+import org.apache.sling.api.servlets.HttpConstants;
 import org.apache.sling.api.servlets.SlingSafeMethodsServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -96,6 +97,21 @@ public class StreamRendererServlet extends SlingSafeMethodsServlet {
             SlingHttpServletResponse response) throws ServletException,
             IOException {
 
+        processRequest(request, response);
+
+    }
+
+    @Override
+    protected void doHead(SlingHttpServletRequest request,
+                         SlingHttpServletResponse response) throws ServletException,
+            IOException {
+        processRequest(request, response);
+    }
+
+    private void processRequest(SlingHttpServletRequest request,
+                                   SlingHttpServletResponse response) throws ServletException,
+            IOException {
+
         // whether this servlet is called as of a request include
         final boolean included = request.getAttribute(SlingConstants.ATTR_REQUEST_SERVLET) != null;
 
@@ -115,11 +131,12 @@ public class StreamRendererServlet extends SlingSafeMethodsServlet {
         }
 
         Resource resource = request.getResource();
+
         if (ResourceUtil.isNonExistingResource(resource)) {
             throw new ResourceNotFoundException("No data to render.");
         }
-
         // trailing slash on url means directory listing
+
         if ("/".equals(request.getRequestPathInfo().getSuffix())) {
             renderDirectory(request, response, included);
             return;
@@ -148,6 +165,11 @@ public class StreamRendererServlet extends SlingSafeMethodsServlet {
         }
         InputStream stream = resource.adaptTo(InputStream.class);
         if (stream != null) {
+            if (isHeadRequest(request)) {
+                setContentLength(response, resource.getResourceMetadata().getContentLength());
+                setHeaders(resource, response);
+                return;
+            }
 
             streamResource(resource, stream, included, request, response);
 
@@ -180,6 +202,10 @@ public class StreamRendererServlet extends SlingSafeMethodsServlet {
     private boolean isRootResourceRequest(Resource resource) {
         return ("/".equals(resource.getPath())) ||
             ("/".equals(resource.getResourceResolver().map(resource.getPath())));
+    }
+
+    private boolean isHeadRequest(HttpServletRequest request) {
+        return HttpConstants.METHOD_HEAD.equals(request.getMethod());
     }
 
     /**
@@ -294,6 +320,12 @@ public class StreamRendererServlet extends SlingSafeMethodsServlet {
             Resource fileRes = resolver.getResource(resource, index);
             if (fileRes != null && !ResourceUtil.isSyntheticResource(fileRes)) {
 
+                setHeaders(fileRes, response);
+
+                if (isHeadRequest(request)) {
+                    return;
+                }
+
                 // include the index resource with no suffix and selectors !
                 RequestDispatcherOptions rdo = new RequestDispatcherOptions();
                 rdo.setReplaceSuffix("");
@@ -307,14 +339,16 @@ public class StreamRendererServlet extends SlingSafeMethodsServlet {
                     dispatcher = request.getRequestDispatcher(fileRes, rdo);
                 }
 
-                setHeaders(fileRes, response);
-
                 dispatcher.include(request, response);
                 return;
             }
         }
 
         if (index) {
+            if (isHeadRequest(request)) {
+                setHeaders(resource, response);
+                return;
+            }
             renderIndex(resource, response);
         } else {
             response.sendError(HttpServletResponse.SC_FORBIDDEN);
